Commenting 13:45 01/26/2023
Commented through all of the code except Main.java. (Main will be rewritten soon)
Added Configuration.java in preparation for population-based training to optimize hyperparameters.

+++ src/Functions.java
@@ -1,8 +1,38 @@
+//-----[IMPORTS]-----\\
+
+//-----[CLASS]-----\\
+/**
+ * Functions
+ * Class
+ * 
+ * Provides static methods for all functions.
+ * 
+ * @author Joshua Savoie
+ */
 public class Functions {
     
+    //-----[SIGMOID]-----\\
+
+    /**
+     * sigmoid()
+     * 
+     * Regular sigmoid.
+     * 
+     * @param x
+     * @return 1 / (1 + e^-x)
+     */
     public static double sigmoid(double x) {
-        return (double)(1 / (1 + (double)Math.pow(Math.E, -x)));
+        return 1 / (1 + Math.pow(Math.E, -x));
     }
+
+    /**
+     * dSigmoid()
+     * 
+     * Derivative of regular sigmoid.
+     * 
+     * @param x
+     * @return sigmoid(x) * (1 - sigmoid(x))
+     */
     public static double dSigmoid(double x) {
         return sigmoid(x) * (1 - sigmoid(x));
     }
